com.unity.ai.generators@1.6.0-pre.1 ## [1.6.0-pre.1] - 2026-02-09
### Changed

- Improved responsiveness of spritesheet generation UI.

### Fixed

- Fixed generated mesh preview lighting under HDRP.

+++ Modules/Unity.AI.Generators.Tools/AssetGeneratorHandle.cs
@@ -41,7 +41,7 @@ namespace Unity.AI.Generators.Tools
         /// <summary>
         /// The estimated point cost for this generation, if available.
         /// </summary>
-        public int PointCost { get; internal set; }
+        public long PointCost { get; internal set; }
 
         internal GenerationHandle(T placeholder, Task<T> generationTask)
         {
